Add wave system
Very basic, but does work, and you can also end the wave by killing all the enemies on the list. Sends you back to the title screen for now.

This also led to me realising that a lot of the manager classes need rewriting to no longer be static, as otherwise resetting them and offloading their data is a PAIN. That'll take a while, but it has already begun as of this commit.