Rework animation system
Deleted AnimationHandler and went back to just calling animations by name. This also means there's no longer randomisation (for now), and for good reason as will be elaborated on in a later update.

Also, finally got rid of the "wrapper" system for AnimationEvents, instead creating a ScriptableObject to store "frame data" that way instead.